@@ -7,7 +7,7 @@
     "require": {
         "php": "^8.1",
         "guzzlehttp/guzzle": "^7.2",
-        "laravel/framework": "^10.10",
+        "laravel/framework": "^11.0",
         "laravel/sanctum": "^3.3",
         "laravel/tinker": "^2.8"
     },
