@@ -1,8 +1,8 @@
 <?php
 
 namespace Database\Seeders;
-use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 
+use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
 
 class AttendeeSeeder extends Seeder
@@ -14,10 +14,12 @@ class AttendeeSeeder extends Seeder
     {
         $users = \App\Models\User::all();
         $events = \App\Models\Event::all();
+
         foreach ($users as $user) {
             $eventsToAttend = $events->random(rand(1, 3));
+
             foreach ($eventsToAttend as $event) {
-                \App\Models\Attendee::created([
+                \App\Models\Attendee::create([
                     'user_id' => $user->id,
                     'event_id' => $event->id
                 ]);
